requirements.txt
streamlit==1.41.1
PyPDF2==3.0.1
pillow==11.0.0